refactor(bcos-evm): fold OpExecCommon into OpTransition; rename executeMessage -> runTxMessage
The shared execution core (the verbatim middle of evmone's transition():
warm access -> build_message + 7702 delegation -> host.call -> refund +
7623 floor) moves from its own OpExecCommon.{h,cpp} pair into
OpTransition.{h,cpp}, keeping a single copy shared by opTransition and
runDeposit. Pure file reorganization -- the copied statements are
unchanged; the function is renamed runTxMessage to pair with runDeposit
(the function boundary is ours, not upstream's, so the rename does not
touch the copy surface).

Verified: standalone bcos-evm build green; BcosEvmEthTests +
BcosEvmOpstackTests (124 cases incl. the 33-vector op-geth differential
gate) 100% pass.
