--- bcos-evm/bcos-evm/opstack/OpExecCommon.cpp
@@ -1,75 +1,0 @@
-#include <bcos-evm/opstack/OpExecCommon.h>
-#include <bcos-evm/opstack/OpHost.h>
-#include <algorithm>
-#include <cassert>
-#include <evmone/delegation.hpp>
-
-namespace bcos::evm::opstack
-{
-namespace
-{
-evmc_message build_message(
-    const evmone::state::Transaction& tx, int64_t execution_gas_limit) noexcept
-{
-    const auto recipient = tx.to.has_value() ? *tx.to : evmc::address{};
-
-    return {
-        .kind = tx.to.has_value() ? EVMC_CALL : EVMC_CREATE,
-        .flags = 0,
-        .depth = 0,
-        .gas = execution_gas_limit,
-        .recipient = recipient,
-        .sender = tx.sender,
-        .input_data = tx.data.data(),
-        .input_size = tx.data.size(),
-        .value = intx::be::store<evmc::uint256be>(tx.value),
-        .create2_salt = {},
-        .code_address = recipient,
-        .code = nullptr,
-        .code_size = 0,
-    };
-}
-}  // namespace
-
-ExecOutcome executeMessage(evmone::state::State& state, OpHost& host,
-    const evmone::state::Transaction& tx, evmc_revision rev, const evmc::address& coinbase,
-    int64_t execution_gas_limit, int64_t min_gas_cost, int64_t delegation_refund)
-{
-    state.get(tx.sender).access_status = EVMC_ACCESS_WARM;
-    if (tx.to.has_value())
-        host.access_account(*tx.to);
-    for (const auto& [a, storage_keys] : tx.access_list)
-    {
-        host.access_account(a);
-        for (const auto& key : storage_keys)
-            state.get_storage(a, key).access_status = EVMC_ACCESS_WARM;
-    }
-    if (rev >= EVMC_SHANGHAI)
-        host.access_account(coinbase);
-
-    auto message = build_message(tx, execution_gas_limit);
-    if (tx.to.has_value())
-    {
-        if (const auto delegate = evmone::get_delegate_address(host, *tx.to))
-        {
-            message.code_address = *delegate;
-            message.flags |= EVMC_DELEGATED;
-            host.access_account(message.code_address);
-        }
-    }
-
-    auto result = host.call(message);
-
-    auto gas_used = tx.gas_limit - result.gas_left;
-
-    const auto max_refund_quotient = rev >= EVMC_LONDON ? 5 : 2;
-    const auto refund_limit = gas_used / max_refund_quotient;
-    const auto refund = std::min(delegation_refund + result.gas_refund, refund_limit);
-    gas_used -= refund;
-    assert(gas_used > 0);
-
-    gas_used = std::max(gas_used, min_gas_cost);
-
-    return {std::move(result), gas_used};
-}
-}  // namespace bcos::evm::opstack
